Move EKumiPlugin to the fr.kazejiyu.ekumi package
The core package is dedicated to workflow definition, it is hence
not an appropriate place for the EKumi plug-in's activator.
